Fix(CommGui): Fix some compilation warning.

+++ src/server/comm/guicommserver.cpp
@@ -67,7 +67,7 @@ CommString GuiCommChannel::readMessage() {
     return message;
 }
 
-bool GuiCommChannel::containsValidJson(const CommString &message, int &endIndex) const {
+bool GuiCommChannel::containsValidJson(const CommString &message, size_t &endIndex) const {
     endIndex = -1;
     if (message.empty() || (message[0] != '{' && message[0] != '[')) {
         return false;
